Remove submodules since they can be published on their own
Per GitHub Pages docs [here], any repo using GitHub Pages will
automatically publish at the configured custom domain (from the user/org
level) using the repo name as its path. This means that the contents
of this code-of-conduct repo can be automatically found at
bridgefoundry.org/code-of-conduct without needing to pull it in as a
submodule. Same for `workshop-map`.

[here]: https://docs.github.com/en/pages/configuring-a-custom-domain-for-your-github-pages-site/about-custom-domains-and-github-pages#supported-custom-domains